Markdown doc specifies that list item may have paragraphs if those are indented by 4 spaces (or a tab) from current list item marker indentation (paragraph meaning there is empty line in between).
If there is empty line and next line is not indented 4 chars then
that should end the above list. This doesn't happen in all markdown
implementations.

In our NEWS case this problem exists in release 0.6 documentation.

It can be avoided by removing 2 leading spaces in lines that are not
list items and requiring all that indents are 0, 2, and 4+ (to make
regexp below work).

Nested lists are supported but one needs to be more careful with
markup there (as the hack below works only on first level).

( The regexp in newssplit.pl mentioned above is
	s/^[ ][ ]// unless /^[ ][ ](?:[\s*+-]|\d+\.)\s/; ).