emacstips: update tagging lisp snippets
Update the tagging snippets to reflect the latest version of notmuch
tag. In fact these snippets work back to at least 0.13 (as lists of
tag changes were always allowed) so this simplifies the page
considerably. We deem pre 0.13 as archaic (debian stable has 0.13) and
remove the examples for older notmuches.